working footers
added two footers depending on the situation. The main header should be used when there are products on the page. The second should be used when there are none. The footer does not work as intended but I do not have the time to fix or worry about it now.